1、新增数据权限注解，实现自动添加过滤查询条件 3、调整部门表、角色表 4、新增工具类

+++ back-end/color-system/src/main/java/com/colorone/system/service/impl/BaseRoleServiceImpl.java
@@ -37,11 +37,11 @@ public class BaseRoleServiceImpl implements BaseRoleService {
     public Integer addBaseRole(BaseRole role) {
         int i = baseRoleMapper.insert(role);
         if (i > 0 && role.getParams().containsKey("permits")) {
-            List<Long> permits = (List<Long>) role.getParams().get("permits");
-            for (Long p : permits) {
+            List<Integer> permits = (List<Integer>) role.getParams().get("permits");
+            for (Integer p : permits) {
                 BaseRoleMenu roleMenu = new BaseRoleMenu();
                 roleMenu.setRoleId(role.getRoleId());
-                roleMenu.setMenuId(p);
+                roleMenu.setMenuId(p.longValue());
                 i += baseRoleMenuMapper.insert(roleMenu);
             }
         }
